src/dcostokenfetcher/requirements.txt
pyinstaller
click
requests
datetime
pyjwt
staticx
patchelf-wrapper
